fix(build): delay() takes Long — coerceAtLeast(Int) returns Int
The 50 ms sampling loop passed an Int to kotlinx.coroutines.delay;
coerceAtLeast(10) keeps the floor at 10 ms and toLong() satisfies the
overload.

+++ app/src/main/java/com/omnix/assistant/presentation/core/OmnixCore.kt
@@ -181,7 +181,7 @@ fun OmnixCore(
         if (state.isAudioReactive && !reduced) {
             while (true) {
                 amplitudeHistory.push(currentAudioLevel)
-                delay(motion.amplitudeSampleMs.coerceAtLeast(10))
+                delay(motion.amplitudeSampleMs.coerceAtLeast(10).toLong())
             }
         }
     }
